refactor: reorganizing imports and fixing file io and plotting code to align with new refactors.
